Updated utility classes file structure - Moved `Weak` collection classes to `Utils.Generics` - Moved `Utils.ImprovedInputHandler` to `Utils.Compatibility.ImprovedInputHandler` - Added doc comments for some new classes - Fixed mod breaking if IIC/IIC:E is enabled (due to not registering keybinds at the right time)
