+++ ControlLib/src/Utils/CompatibilityManager.cs
@@ -2,10 +2,27 @@ using System.Linq;
 
 namespace ControlLib.Utils;
 
+/// <summary>
+/// Simple helper for determining the presence of other mods and ensure mod compatibility.
+/// </summary>
 public static class CompatibilityManager
 {
+    /// <summary>
+    /// Determines if a given mod is currently enabled.
+    /// </summary>
+    /// <param name="modID">The ID of the mod to check for.</param>
+    /// <returns><c>true</c> if the given mod was found to be enabled, <c>false</c> otherwise.</returns>
     public static bool IsModEnabled(string modID) => ModManager.ActiveMods.Any(mod => mod.id == modID);
 
+    /// <summary>
+    /// Determines if either Improved Input Config or Improved Input Config: Extended are enabled.
+    /// </summary>
+    /// <returns><c>true</c> if one of these mods is enabled, <c>false</c> otherwise.</returns>
     public static bool IsIICEnabled() => IsModEnabled("improved-input-config");
+
+    /// <summary>
+    /// Determines if the Rain Meadow mod is enabled.
+    /// </summary>
+    /// <returns><c>true</c> if the mod is enabled, <c>false</c> otherwise.</returns>
     public static bool IsRainMeadowEnabled() => IsModEnabled("henpemaz_rainmeadow");
 }
